backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sophub-backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic-settings",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "sqlalchemy",
    "psycopg[binary]",
    "alembic",
    "python-multipart",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "httpx",
    "pytest-asyncio",
]

[tool.uvicorn]
factory = true
app = "app.main:create_app"
host = "127.0.0.1"
port = 8000
reload = true


